A plain text test should be in plain text formatter tests
